+++ src/components/Unscheduled.tsx
@@ -7,10 +7,10 @@ import { memo } from 'react'
 import { useCollapsed } from 'src/services/util'
 import Button from './Button'
 
-const Unscheduled = memo(_Unscheduled, _.isEqual)
+const Unscheduled = memo(_Unscheduled, () => true)
 export default Unscheduled
 
-function _Unscheduled({ isFlex }: { isFlex: boolean }) {
+function _Unscheduled() {
   const showCompleted = useAppStore((state) => state.settings.showCompleted)
   const tasks = useAppStore((state) =>
     _.filter(
@@ -19,24 +19,38 @@ function _Unscheduled({ isFlex }: { isFlex: boolean }) {
         !task.scheduled && (!task.completed || showCompleted) && !task.parent
     )
   )
-  const calendarMode = useAppStore((state) => state.calendarMode)
+  const childWidth = useAppStore((state) =>
+    (state.viewMode === 'week' || state.viewMode === 'hour') &&
+    state.childWidth > 1
+      ? state.childWidth
+      : 1
+  )
+
   const { collapsed, allHeadings } = useCollapsed(tasks)
   return (
-    <div className={`h-0 grow flex flex-col ${calendarMode ? '!w-full' : ''}`}>
-      <div className='flex items-center space-x-1 group'>
+    <div className={`h-0 grow flex flex-col`}>
+      <div className='flex items-center space-x-1 group flex-none'>
         <Button
           className='w-6 opacity-0 group-hover:opacity-100 transition-opacity duration-300'
           onClick={() => setters.patchCollapsed(allHeadings, !collapsed)}
           src={collapsed ? 'chevron-right' : 'chevron-down'}
         />
         <Droppable id='unscheduled-timespan' data={{ scheduled: '' }}>
-          <div className='font-menu mb-2'>{'Unscheduled'}</div>
+          <div className='font-menu'>{'Unscheduled'}</div>
         </Droppable>
       </div>
       <div
-        className={`h-0 grow rounded-lg bg-secondary-alt ${
-          isFlex
-            ? 'overflow-hidden child:flex child:h-full child:overflow-x-auto child:snap-x child:snap-mandatory child:child:h-full child:child:overflow-y-auto child:child:flex-none child:child:w-full sm:child:child:w-1/2 lg:child:child:w-1/3 xl:child:child:w-1/4 child:child:snap-start'
+        className={`h-0 grow w-full p-1 rounded-lg bg-secondary-alt mt-1 ${
+          childWidth > 1
+            ? `child:flex child:overflow-y-hidden child:overflow-x-auto child:flex-col child:flex-wrap child:h-full child:child:max-h-full child:child:overflow-y-auto ${
+                [
+                  '',
+                  'child:child:w-full',
+                  'child:child:w-1/2',
+                  'child:child:w-1/3',
+                  'child:child:w-1/4',
+                ][childWidth]
+              }`
             : 'overflow-x-hidden overflow-y-auto'
         }`}
         data-auto-scroll={'y'}