+++ CHANGELOG.md
@@ -18,8 +18,16 @@
 
 ## 1.7.1 (Upcoming)
 
+**Added:**
+- Three layouts now available: One (single date with split day/hours), Row (rolling view of days), and Grid (week view)
+- Day start setting now sets when days transition, allowing days to extend past 12 AM
+- Use Daily Note template when creating new daily notes
+- Improved [borders between dates](https://github.com/joshuatazrein/obsidian-time-ruler/issues/88#issuecomment-1846164954) and button layout for grid view
+- Option to turn off times globally
+
 **Fixed:**
 - Parsing error with daily notes
+- New tasks in notes with headings create them before the first heading, unless one is selected
 
 ## 1.7.0 (12/5/2023)